Enable dnsmasq restart only when ipsets change

+++ src/github.com/zededa/go-provision/zedrouter/zedrouter.go
@@ -153,13 +153,13 @@ func handleInit(configFilename string, statusFilename string,
 		addrs6, err := netlink.AddrList(link, netlink.FAMILY_V6)
 		fmt.Printf("UplinkAddrs v4 %d v6 %d\n", len(addrs4), len(addrs6))
 		globalStatus.UplinkAddrs = make([]net.IP,
-			len(addrs4) + len(addrs6))
+			len(addrs4)+len(addrs6))
 		for i, addr := range addrs4 {
 			globalStatus.UplinkAddrs[i] = addr.IP
 		}
 		for i, addr := range addrs6 {
-			// XXX check if link-local?		
-			globalStatus.UplinkAddrs[i + len(addrs4)] = addr.IP
+			// XXX check if link-local?
+			globalStatus.UplinkAddrs[i+len(addrs4)] = addr.IP
 		}
 	}
 	globalStatus.Uplink = globalConfig.Uplink
@@ -445,7 +445,7 @@ func handleCreate(statusFilename string, configArg interface{}) {
 	}
 
 	ipsets := compileAppInstanceIpsets(config.OverlayNetworkList,
-					config.UnderlayNetworkList)
+		config.UnderlayNetworkList)
 
 	for i, olConfig := range config.OverlayNetworkList {
 		olNum := i + 1
@@ -728,10 +728,10 @@ func handleModify(statusFilename string, configArg interface{},
 		return
 	}
 
-	newIpsets, staleIpsets := updateAppInstanceIpsets(config.OverlayNetworkList,
-							    config.UnderlayNetworkList,
-							    status.OverlayNetworkList,
-							    status.UnderlayNetworkList)
+	newIpsets, staleIpsets, restartDnsmasq := updateAppInstanceIpsets(config.OverlayNetworkList,
+		config.UnderlayNetworkList,
+		status.OverlayNetworkList,
+		status.UnderlayNetworkList)
 
 	// Look for ACL and NametoEidList changes in overlay
 	for i, olConfig := range config.OverlayNetworkList {
@@ -760,18 +760,20 @@ func handleModify(statusFilename string, configArg interface{},
 		// we need to restart dnsmasq (and update its ipset configs?
 		// get a return value from updateAclConfiglet to indicate
 		// whether there were such changes?
-		cfgFilename := "dnsmasq." + olIfname + ".conf"
-		cfgPathname := runDirname + "/" + cfgFilename
-		EID := olConfig.EID
-		olMac := "00:16:3e:1:" + strconv.FormatInt(int64(olNum), 16) +
-		":" + strconv.FormatInt(int64(appNum), 16)
-		stopDnsmasq(cfgFilename, false)
-		//remove old dnsmasq configuration file
-		os.Remove(cfgPathname)
-		createDnsmasqOverlayConfiglet(cfgPathname, olIfname, olAddr1,
-		EID.String(), olMac, hostsDirpath,
-		config.UUIDandVersion.UUID.String(), newIpsets)
-		startDnsmasq(cfgPathname)
+		if restartDnsmasq {
+			cfgFilename := "dnsmasq." + olIfname + ".conf"
+			cfgPathname := runDirname + "/" + cfgFilename
+			EID := olConfig.EID
+			olMac := "00:16:3e:1:" + strconv.FormatInt(int64(olNum), 16) +
+				":" + strconv.FormatInt(int64(appNum), 16)
+			stopDnsmasq(cfgFilename, false)
+			//remove old dnsmasq configuration file
+			os.Remove(cfgPathname)
+			createDnsmasqOverlayConfiglet(cfgPathname, olIfname, olAddr1,
+				EID.String(), olMac, hostsDirpath,
+				config.UUIDandVersion.UUID.String(), newIpsets)
+			startDnsmasq(cfgPathname)
+		}
 
 		// Format a json string with any additional info
 		addInfoApp := types.AdditionalInfoApp{
@@ -812,33 +814,35 @@ func handleModify(statusFilename string, configArg interface{},
 		updateACLConfiglet(ulIfname, false, ulStatus.ACLs,
 			ulConfig.ACLs, 4, "")
 
-		//update underlay dnsmasq configuration
-		cfgFilename := "dnsmasq." + ulIfname + ".conf"
-		cfgPathname := runDirname + "/" + cfgFilename
-		ulAddr1 := "172.27." + strconv.Itoa(appNum) + ".1"
-		ulAddr2 := "172.27." + strconv.Itoa(appNum) + ".2"
-		ulMac := "00:16:3e:0:0:" + strconv.FormatInt(int64(appNum), 16)
-		stopDnsmasq(cfgFilename, false)
-		//remove old dnsmasq configuration file
-		os.Remove(cfgPathname)
-		createDnsmasqUnderlayConfiglet(cfgPathname, ulIfname, ulAddr1,
-		ulAddr2, ulMac,
-		config.UUIDandVersion.UUID.String(), newIpsets)
-		startDnsmasq(cfgPathname)
+		if restartDnsmasq {
+			//update underlay dnsmasq configuration
+			cfgFilename := "dnsmasq." + ulIfname + ".conf"
+			cfgPathname := runDirname + "/" + cfgFilename
+			ulAddr1 := "172.27." + strconv.Itoa(appNum) + ".1"
+			ulAddr2 := "172.27." + strconv.Itoa(appNum) + ".2"
+			ulMac := "00:16:3e:0:0:" + strconv.FormatInt(int64(appNum), 16)
+			stopDnsmasq(cfgFilename, false)
+			//remove old dnsmasq configuration file
+			os.Remove(cfgPathname)
+			createDnsmasqUnderlayConfiglet(cfgPathname, ulIfname, ulAddr1,
+				ulAddr2, ulMac,
+				config.UUIDandVersion.UUID.String(), newIpsets)
+			startDnsmasq(cfgPathname)
+		}
 	}
 
 	// Remove stale ipsets
 	// In case if there are any references to these ipsets from other
 	// domUs kernel would not remove it. ipset destroy command would just fail.
 	for _, ipset := range staleIpsets {
-	    err := ipsetDestroy(fmt.Sprintf("ipv4.%s", ipset))
-	    if err != nil {
-		log.Println("ipset destroy ipv4", ipset, err)
-	    }
-	    err = ipsetDestroy(fmt.Sprintf("ipv6.%s", ipset))
-	    if err != nil {
-		log.Println("ipset destroy ipv6", ipset, err)
-	    }
+		err := ipsetDestroy(fmt.Sprintf("ipv4.%s", ipset))
+		if err != nil {
+			log.Println("ipset destroy ipv4", ipset, err)
+		}
+		err = ipsetDestroy(fmt.Sprintf("ipv6.%s", ipset))
+		if err != nil {
+			log.Println("ipset destroy ipv6", ipset, err)
+		}
 	}
 
 	// Write out what we modified to AppNetworkStatus
